% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_line_chart}
\alias{all_line_chart}
\title{all_line_chart}
\usage{
all_line_chart(data, output_dir, ...)
}
\arguments{
\item{data}{List from data_processer}

\item{output_dir}{Directory to save figures to}
}
\description{
Saves all line charts in list from data_processer
}
