% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart_plot.R
\name{bar_chart_plot}
\alias{bar_chart_plot}
\title{bar_chart_plot}
\usage{
bar_chart_plot(data, output_dir, fill_col = attributes(data)$fill,
  colors = attributes(data)$colors, diff = F, region = NULL, width = 10,
  height = 7, break_size = 10, remove_title = F)
}
\arguments{
\item{data}{Dataframe/tibble of processed GCAM data}

\item{output_dir}{Output directory to save charts to}

\item{fill_col}{Column name, in quotes, to color by. Defaults to attribute of "fill" for processed data.}

\item{colors}{Vector of colors for chart. Defaults to attribute of "colors" for processed data}

\item{diff}{If true, adds subtitle to plot and DIFF to output name}

\item{region}{String of region name to filter by}

\item{break_size}{Interval of years to print on x axis, starting with minimum year in data. Default is 10 years}
}
\description{
Create bar chart of processed data
}
\examples{
bar_chart_plot(df, output_dir = "/outputs", fill = "technology", colors = c("red", "blue"))
bar_chart_plot(data[["primary_energy"]], output_dir = "/outputs")
}
\keyword{bar}
\keyword{chart}
