% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_fe_data.R
\name{building_fe_data}
\alias{building_fe_data}
\title{building_fe_data}
\usage{
building_fe_data(query, scenarios, diff = NULL)
}
\arguments{
\item{query}{GCAM query containing building final energy output data of one or multiple scenarios}

\item{scenario}{GCAM scenarios to include in processed data}
}
\description{
Process building final energy query
}
\examples{
building_fe_data("queryA.csv", c("Reference1,date=2017-9-6T13:43:53-07:00", "Reference2,date=2017-9-6T13:43:53-07:00"))
}
\keyword{building}
\keyword{energy}
\keyword{final}
