% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_data.R
\name{climate_forcing_data}
\alias{climate_forcing_data}
\title{climate_forcing_data}
\usage{
climate_forcing_data(query, scenarios, query_dir = QUERY_FOLDER)
}
\arguments{
\item{query}{GCAM query containing climate forcing data of one or multiple scenarios}

\item{scenario}{GCAM scenarios to include in processed data}
}
\description{
Process climate forcing query
}
\examples{
climate_forcing_data("queryA.csv", c("Reference1,date=2017-9-6T13:43:53-07:00", "Reference2,date=2017-9-6T13:43:53-07:00"))
}
\keyword{climate}
\keyword{forcing}
