% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_data.R
\name{CO2_concentration_data}
\alias{CO2_concentration_data}
\title{CO2_concentration_data}
\usage{
CO2_concentration_data(query, scenarios, query_dir = QUERY_FOLDER)
}
\arguments{
\item{query}{GCAM query containing CO2 concentration data of one or multiple scenarios}

\item{scenario}{GCAM scenarios to include in processed data}
}
\description{
Process CO2 concentration query
}
\examples{
CO2_concentration_data("queryA.csv", c("Reference1,date=2017-9-6T13:43:53-07:00", "Reference2,date=2017-9-6T13:43:53-07:00"))
}
\keyword{CO2}
\keyword{concentration}
