% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_data.R
\name{CO2_emissions_agg_sector_data}
\alias{CO2_emissions_agg_sector_data}
\title{CO2_emissions_agg_sector_data}
\usage{
CO2_emissions_agg_sector_data(query, scenarios, query_dir = QUERY_FOLDER)
}
\arguments{
\item{query}{GCAM query containing CO2 emissions data of one or multiple scenarios}

\item{scenario}{GCAM scenarios to include in processed data}
}
\description{
Process CO2 emissions by aggregate sector query. Not currently mapped
}
\examples{
CO2_emissions_agg_sector_data("queryA.csv", c("Reference1", "Reference2"))
}
\keyword{CO2}
\keyword{emissions}
