% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_processer}
\alias{data_processer}
\title{data_processer}
\usage{
data_processer(folder = QUERY_FOLDER, scenarios = SCENARIOS)
}
\arguments{
\item{folder}{Directory with GCAM queries only}

\item{scenarios}{Vector of scenarios to include in data}
}
\description{
Returns a list of processed data for all queries that have processing code
}
