% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_processer.R
\name{diff_processer}
\alias{diff_processer}
\title{diff_processer}
\usage{
diff_processer(folder = QUERY_FOLDER, scenarios = SCENARIOS,
  diff_scenario = DIFF_SCENARIO)
}
\arguments{
\item{folder}{Directory with GCAM queries only}

\item{scenarios}{Vector of scenarios to include in data}

\item{diff_scenario}{Scenario to diff all other scenarios by}
}
\description{
Returns a list of processed data for all queries that have processing code, showing difference between scenarios and a reference scenario
}
