% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elec_gen_data.R
\name{elec_gen_data}
\alias{elec_gen_data}
\title{elec_gen_data}
\usage{
elec_gen_data(query, scenarios, query_dir = QUERY_FOLDER)
}
\arguments{
\item{query}{GCAM query containing primary energy data of one or multiple scenarios}

\item{scenario}{GCAM scenarios to include in processed data}
}
\description{
Process electricity generation query
}
\examples{
elec_gen_data("queryA.csv", c("Reference1,date=2017-9-6T13:43:53-07:00", "Reference2,date=2017-9-6T13:43:53-07:00"))
}
\keyword{electricity}
\keyword{generation}
