% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gas_prod_data.R
\name{gas_prod_data}
\alias{gas_prod_data}
\title{gas_prod_data}
\usage{
gas_prod_data(query, scenarios, query_dir = QUERY_FOLDER)
}
\arguments{
\item{query}{GCAM query containing gas production output data of one or multiple scenarios}

\item{scenario}{GCAM scenarios to include in processed data}
}
\description{
Process gas production query
}
\examples{
gas_prod_data("queryA.csv", c("Reference1,date=2017-9-6T13:43:53-07:00", "Reference2,date=2017-9-6T13:43:53-07:00"))
}
\keyword{gas}
\keyword{production}
