% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/industry_fe.data.R
\name{industry_fe_data}
\alias{industry_fe_data}
\title{industry_fe_data}
\usage{
industry_fe_data(query, scenarios, query_dir = QUERY_FOLDER)
}
\arguments{
\item{query}{GCAM query containing industry final energy output data of one or multiple scenarios}

\item{scenario}{GCAM scenarios to include in processed data}
}
\description{
Process industry final energy query
}
\examples{
industry_fe_data("queryA.csv", c("Reference1,date=2017-9-6T13:43:53-07:00", "Reference2,date=2017-9-6T13:43:53-07:00"))
}
\keyword{energy}
\keyword{final}
\keyword{industry}
