% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/industry_so_enuse_data.R
\name{industry_so_enuse_data}
\alias{industry_so_enuse_data}
\title{industry_so_enuse_data}
\usage{
industry_so_enuse_data(query, scenarios, query_dir = QUERY_FOLDER)
}
\arguments{
\item{query}{GCAM query containing building final energy output data of one or multiple scenarios}

\item{scenario}{GCAM scenarios to include in processed data}
}
\description{
Process industrial primary output query for only industrial energy use
}
\examples{
industry_so_enuse_data("queryA.csv", c("Reference1,date=2017-9-6T13:43:53-07:00", "Reference2,date=2017-9-6T13:43:53-07:00"))
}
\keyword{building}
\keyword{energy}
\keyword{final}
