% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_alloc_data.R
\name{land_alloc_data}
\alias{land_alloc_data}
\title{land_alloc_data}
\usage{
land_alloc_data(query, scenarios, query_dir = QUERY_FOLDER)
}
\arguments{
\item{query}{GCAM query containing land allocation data of one or multiple scenarios}

\item{scenario}{GCAM scenarios to include in processed data}
}
\description{
Process land allocation query
}
\examples{
land_alloc_data("queryA.csv", c("Reference1,date=2017-9-6T13:43:53-07:00", "Reference2,date=2017-9-6T13:43:53-07:00"))
}
\keyword{allocation}
\keyword{land}
