% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_chart_plot.R
\name{line_chart_plot}
\alias{line_chart_plot}
\title{line_chart_plot}
\usage{
line_chart_plot(data, output_dir, fill_col = "scenario", diff = F,
  region = NULL, width = 10, height = 7, break_size = 10,
  manual_colors = NULL, jitter = F)
}
\arguments{
\item{data}{Dataframe/tibble of processed GCAM data}

\item{output_dir}{Output directory to save charts to}

\item{fill_col}{Column name, in quotes, to color by. Default is "scenario". To use other column, may need to add facet.}

\item{diff}{If true, adds subtitle to plot and DIFF to output name}

\item{region}{String of region name to filter by}

\item{break_size}{Interval of years to print on x axis, starting with minimum year in data. Default is 10 years}
}
\description{
Create line chart of processed data
}
\examples{
line_chart_plot(df, output_dir = "/outputs")
line_chart_plot(data[["primary_energy"]], output_dir = "/outputs")
}
\keyword{chart}
\keyword{line}
