% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_transport_so_data.R
\name{pass_transport_so_data}
\alias{pass_transport_so_data}
\title{pass_transport_so_data}
\usage{
pass_transport_so_data(query, scenarios)
}
\arguments{
\item{query}{GCAM query containing transportation service output data of one or multiple scenarios}

\item{scenario}{GCAM scenarios to include in processed data}
}
\description{
Process transportation service output query - filtering out only trn_pass_road_LDV_4W
}
\examples{
pass_transport_so_data("queryA.csv", c("Reference1,date=2017-9-6T13:43:53-07:00", "Reference2,date=2017-9-6T13:43:53-07:00"))
}
\keyword{output}
\keyword{passenger}
\keyword{service}
\keyword{transportation}
