% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_energy_data.R
\name{primary_energy_data}
\alias{primary_energy_data}
\title{primary_energy_data}
\usage{
primary_energy_data(query, scenarios)
}
\arguments{
\item{query}{GCAM query containing primary energy data of one or multiple scenarios}

\item{scenario}{GCAM scenarios to include in processed data}
}
\description{
Process primary energy query
}
\examples{
primary_energy_data("queryA.csv", c("Reference1,date=2017-9-6T13:43:53-07:00", "Reference2,date=2017-9-6T13:43:53-07:00"))
}
\keyword{energy}
\keyword{primary}
