% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_id.R
\name{query_id}
\alias{query_id}
\title{query_id}
\usage{
query_id(folder)
}
\arguments{
\item{folder}{Directory with GCAM queries only}
}
\description{
Creates tibble with query output file names and query names
}
\examples{
query_id(output_dir = "gcam_queries")
}
