% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{query_splitter}
\alias{query_splitter}
\title{query_splitter}
\usage{
query_splitter(query_file, output_dir)
}
\arguments{
\item{query_file}{Query file location}

\item{output_dir}{Where output csvs are to be written}
}
\description{
For one csv file with multiple queries. Writes into individual query csv files
}
