% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refined_liquids_data.R
\name{refined_liquids_data}
\alias{refined_liquids_data}
\title{refined_liquids_data}
\usage{
refined_liquids_data(query, scenarios, query_dir = QUERY_FOLDER)
}
\arguments{
\item{query}{GCAM query containing refined liquids data of one or multiple scenarios}

\item{scenario}{GCAM scenarios to include in processed data}
}
\description{
Process refined liquids query
}
\examples{
refined_liquids_data("queryA.csv", c("Reference1,date=2017-9-6T13:43:53-07:00", "Reference2,date=2017-9-6T13:43:53-07:00"))
}
\keyword{liquids}
\keyword{refined}
