% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scenario_rename}
\alias{scenario_rename}
\title{scenario_rename}
\usage{
scenario_rename(data, name_lookup)
}
\arguments{
\item{data}{Processed GCAm data}

\item{name_lookup}{Vector of old scenarios (without timestamp) and new names, see example}
}
\description{
Returns a list of processed data for all queries that have processing code
}
\examples{
scenario_rename(data = df[["primary_energy"]], name_lookup = c(Reference = "BAU", Ctax20 = "Carbon Tax $20"))
lapply(data, scenario_rename, name_lookup = c(Reference = "BAU", Ctax20 = "Carbon Tax $20"))
}
