% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornado_chart_plot.R
\name{tornado_chart_plot}
\alias{tornado_chart_plot}
\title{tornado_chart_plot}
\usage{
tornado_chart_plot(data, output_dir, ref_scenario, savename,
  plot_title = NULL, color_flip = F, region = NULL, width = 10,
  height = 7, break_size = 10)
}
\arguments{
\item{data}{Dataframe/tibble of processed GCAM data}

\item{output_dir}{Output directory to save charts to}

\item{ref_scenario}{Scenario to use as base for tornado to start from}

\item{savename}{Name to save plot under}

\item{plot_title}{Title to add, defaults to no title}

\item{color_flip}{Default is green above, red below. Switch to T to flip}

\item{region}{String of region name to filter by}

\item{break_size}{Interval of years to print on x axis, starting with minimum year in data. Default is 10 years}
}
\description{
Create tornado chart of one year/variable from processed data
}
\keyword{bar}
\keyword{chart}
