% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_fe_data.R
\name{transport_fe_data}
\alias{transport_fe_data}
\title{transport_fe_data}
\usage{
transport_fe_data(query, scenarios, query_dir = QUERY_FOLDER)
}
\arguments{
\item{query}{GCAM query containing transport final energy output data of one or multiple scenarios}

\item{scenario}{GCAM scenarios to include in processed data}
}
\description{
Process transport final energy query
}
\examples{
transport_fe_data("queryA.csv", c("Reference1,date=2017-9-6T13:43:53-07:00", "Reference2,date=2017-9-6T13:43:53-07:00"))
}
\keyword{energy}
\keyword{final}
\keyword{transport}
